% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf.R
\docType{methods}
\name{as.data.frame,DDF-method}
\alias{as.data.frame,DDF-method}
\title{Return all the of a DistributedDataFrame as a R datafrme}
\usage{
\S4method{as.data.frame}{DDF}(x)
}
\arguments{
\item{x}{a DistributedDataFrame}
}
\value{
an R native data.frame
}
\description{
Return all the of a DistributedDataFrame as a R datafrme
}
\details{
\code{as.data.frame} for a DistributedDataFrame returns all the data of that DistributedDataFrame as an R native data.frame.
}

