% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.DDF,data.frame-method}
\alias{as.DDF,data.frame-method}
\title{Convert a data.frame to a DDF}
\usage{
\S4method{as.DDF}{data.frame}(x, dm)
}
\arguments{
\item{dm}{a DDF manager}
}
\description{
Convert a data.frame to a DDF
}

