% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf.R
\docType{methods}
\name{colnames,DDF-method}
\alias{colnames,DDF-method}
\title{Retrieve a DistributedDataFrame's column names}
\usage{
\S4method{colnames}{DDF}(x)
}
\arguments{
\item{x}{a DDF object}
}
\value{
a character vector
}
\description{
Retrieve a DistributedDataFrame's column names
}

