% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataFrame}
\alias{dataFrame}
\title{Return all the of a DistributedDataFrame as a R datafrme}
\usage{
dataFrame(x, ...)
}
\arguments{
\item{x}{a DistributedDataFrame}
}
\value{
an R native data.frame
}
\description{
Return all the of a DistributedDataFrame as a R datafrme
}
\details{
\code{dataFrame} for a DistributedDataFrame returns all the data of that DistributedDataFrame as an R native data.frame.
}

