% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ddfKmeans}
\alias{ddfKmeans}
\title{ML model KMeans}
\usage{
ddfKmeans(x, ...)
}
\arguments{
\item{x}{a Distributed Data Frame.}

\item{centers}{number of cluster to be clusterd.}

\item{runs}{number of runs.}

\item{maxIters}{max times of iterations.}
}
\value{
a IModel of distributedDataFrame for KMeans classification
}
\description{
Return Returns IModel of distributedDataFrame
}

