% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-manager.R
\docType{class}
\name{DDFManager-class}
\alias{DDFManager}
\alias{DDFManager-class}
\title{An S4 class that represents a DistributedDataFrame Manager}
\usage{
DDFManager(engine = c("spark", "flink", "jdbc", "sfdc", "postgres", "aws",
  "redshift", "basic"))
}
\arguments{
\item{engine}{the underlying engine, for example: spark or local}
}
\description{
An S4 class that represents a DistributedDataFrame Manager
}

