% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fivenum,DDF-method}
\alias{fivenum,DDF-method}
\title{Tukey Five-Number Summaries}
\usage{
\S4method{fivenum}{DDF}(x)
}
\arguments{
\item{x}{a Distributed Data Frame.}
}
\value{
a data.frame in which each column is a vector containing the summary information.
}
\description{
Return Returns Tukey's five number summary (minimum, lower-hinge, median, upper-hinge, maximum)
for each numeric column of a Distributed Data Frame.
}

