% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\docType{class}
\name{GeneralizedLinearModel-class}
\alias{GeneralizedLinearModel-class}
\title{A S4 class that represents a GeneralizedLinearModel model}
\arguments{
\item{jmlmodel}{Java object reference to the backing MLlib model}
}
\description{
A S4 class that represents a GeneralizedLinearModel model
}
\section{Slots}{

\describe{
\item{\code{centers}}{a matrix of centers}
}}

