% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf.R
\docType{methods}
\name{head,DDF-method}
\alias{head,DDF-method}
\title{Return the First Part of a DistributedDataFrame}
\usage{
\S4method{head}{DDF}(x, n = 6L)
}
\arguments{
\item{x}{a DistributedDataFrame}

\item{n}{a single positive integer, rows for the resulting \code{data.frame}.}
}
\value{
an R native data.frame
}
\description{
Return the First Part of a DistributedDataFrame
}
\details{
\code{head} for a DistributedDataFrame returns the first rows of that DistributedDataFrame as an R native data.frame.
}

