% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{KmeansModel-class}
\alias{KmeansModel-class}
\title{A S4 class that represents a KMeans model}
\arguments{
\item{jmlmodel}{Java object reference to the backing MLlib model}
}
\description{
A S4 class that represents a KMeans model
}
\section{Slots}{

\describe{
\item{\code{centers}}{a matrix of centers}
}}

