% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-manager.R
\name{load_file}
\alias{load_file}
\title{Load local file}
\usage{
load_file(x, ...)
}
\arguments{
\item{x}{a DDFManager object}

\item{path}{path to the local file}

\item{sep}{separator}
}
\value{
a DDF
}
\description{
Load local file
}

