% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-manager.R
\name{load_jdbc}
\alias{load_jdbc}
\title{Load data from JDBC}
\usage{
load_jdbc(x, uri, username, password, table)
}
\arguments{
\item{x}{a DDFManager object}

\item{uri}{jdbc uri including the database name}

\item{username}{username}

\item{password}{password}

\item{table}{table}
}
\value{
a DDF
}
\description{
Load data from JDBC
}

