% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{mean,DDF-method}
\alias{mean,DDF-method}
\title{Calculate Mean value of each DDF Column}
\usage{
\S4method{mean}{DDF}(x)
}
\arguments{
\item{x}{a Distributed Data Frame.}
}
\value{
a vector contains mean values of all columns
}
\description{
Return the mean value of each DDF column.
}

