% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf.R
\docType{methods}
\name{merge,DDF,DDF-method}
\alias{merge,DDF,DDF-method}
\title{Merge two DDFs}
\usage{
\S4method{merge}{DDF,DDF}(x, y, by = intersect(colnames(x), colnames(y)),
  by.x = by, by.y = by, type = c("inner", "left", "right", "full",
  "left.semi"))
}
\description{
Merge two DDFs
}

