% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-ml.R
\docType{class}
\name{MLModel-class}
\alias{MLModel}
\alias{MLModel-class}
\title{A S4 class that represents a ML model of DistributedDataFrame}
\usage{
MLModel(jmlmodel)
}
\arguments{
\item{jmlmodel}{Java object reference to the backing MLlib model}
}
\description{
A S4 class that represents a ML model of DistributedDataFrame
}

