% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{na.omit,DDF-method}
\alias{na.omit,DDF-method}
\title{Drop NA values}
\usage{
\S4method{na.omit}{DDF}(object, by = "ROW", inplace = FALSE)
}
\arguments{
\item{by}{the axis by which to drop if NA value exits, Row represents by Row as default, COLUMN is column.}

\item{inplace}{whether to treat the ddf inplace, default is FALSE.}

\item{x}{a Distributed Data Frame.}
}
\value{
a data.frame with no NA values.
}
\description{
Return a ddf instance with no NAs.
}

