% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ncol,DDF-method}
\alias{ncol,DDF-method}
\title{Get a DistributedDataFrame's number of columns}
\usage{
\S4method{ncol}{DDF}(x)
}
\arguments{
\item{x}{a DDF object}
}
\value{
an integer of length 1 or NULL
}
\description{
Get a DistributedDataFrame's number of columns
}

