% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{nrow,DDF-method}
\alias{nrow,DDF-method}
\title{Get a DistributedDataFrame's number of rows}
\usage{
\S4method{nrow}{DDF}(x)
}
\arguments{
\item{x}{a DDF object}
}
\value{
an integer of length 1 or NULL
}
\description{
Get a DistributedDataFrame's number of rows
}

