% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf.R
\docType{methods}
\name{[-DDF}
\alias{[-DDF}
\title{DistributedDataFrame subsetting/filtering for extracting values}
\usage{
\S4method{[}{DDF}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{the DistributedDataFramee.}

\item{i}{empty or an expression to filter rows.}

\item{j}{column names or indices.}

\item{drop}{coerce the DistributedDataFrame result to a DistributedVector if only one column is specified.}
}
\value{
a DistributedDataFrame or a DistributedVector.
}
\description{
DistributedDataFrame subsetting/filtering for extracting values
}
\section{Usage}{

\describe{
 \code{x[i,j,...,drop=TRUE]}
}
}

