% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sample,DDF-method}
\alias{sample,DDF-method}
\title{Return a sample of specific size from a DistributedDataFrame}
\usage{
\S4method{sample}{DDF}(x, size, replace = FALSE, seed = 123L)
}
\arguments{
\item{x}{a DistributedDataFrame}

\item{size}{number of samples}

\item{replace}{sampling with or without replacement}

\item{seed}{a random seed}
}
\description{
Return a sample of specific size from a DistributedDataFrame
}

