% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf.R
\name{sample2ddf}
\alias{sample2ddf}
\title{Sample a fraction of rows from a DistributedDataFrame}
\usage{
sample2ddf(x, ...)
}
\arguments{
\item{x}{a DistributedDataFrame}

\item{percent}{a double value specify the fraction}

\item{replace}{sampling with or without replacement}

\item{seed}{a random seed}
}
\value{
a DistributedDataFrame
}
\description{
Sample a fraction of rows from a DistributedDataFrame
}

