% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-manager.R
\docType{methods}
\name{sql,DDFManager,character-method}
\alias{sql,DDFManager,character-method}
\title{Execute a SQL and return an R data.frame}
\usage{
\S4method{sql}{DDFManager,character}(x, sql, queryOnDDF = TRUE)
}
\arguments{
\item{x}{a DDFManager object}

\item{sql}{the query string}

\item{queryOnDDF}{whether the query is on ddf, or on the original engine}
}
\value{
an R data.frame
}
\description{
Execute a SQL and return an R data.frame
}

