% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddf-manager.R
\docType{methods}
\name{sql2ddf,DDFManager,character-method}
\alias{sql2ddf,DDFManager,character-method}
\title{Execute a HiveQL and return a DistributedDataFrame}
\usage{
\S4method{sql2ddf}{DDFManager,character}(x, sql, queryOnDDF = TRUE)
}
\arguments{
\item{x}{a DDFManager object}

\item{sql}{the query, only support SELECT ones}
}
\value{
a DistributedDataFrame
}
\description{
Execute a HiveQL and return a DistributedDataFrame
}

