% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,DDF-method}
\alias{summary,DDF-method}
\title{Return a statistical summary of a DistributedDataFrame's columns}
\usage{
\S4method{summary}{DDF}(object)
}
\arguments{
\item{object}{a DistributedDataFrame}
}
\value{
a data.frame containing summary numbers
}
\description{
Return a statistical summary of a DistributedDataFrame's columns
}

