% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{var,DDF-method}
\alias{var,DDF-method}
\title{Calculate the variance and std of a column in DDF}
\usage{
\S4method{var}{DDF}(x, y)
}
\arguments{
\item{x}{a Distributed Data Frame.}

\item{col}{the column name or index of x}
}
\value{
a vector with dimension 2.
}
\description{
Return a vector with dimension 2, 1st dimension denotes the variance and 2ed dimension denotes the standard variance.
}

