% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{colnames,DDF-method}
\alias{colnames,DDF-method}
\title{Retrieve a DistributedDataFrame's column names}
\usage{
\S4method{colnames}{DDF}(x)
}
\arguments{
\item{x}{a DDF object}
}
\value{
a character vector
}
\description{
Retrieve a DistributedDataFrame's column names
}

