% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DDF-class}
\alias{DDF}
\alias{DDF-class}
\title{A S4 class that represents a DistributedDataFrame}
\usage{
DDF(jddf)
}
\arguments{
\item{jddf}{Java object reference to the backing DistributedDataFrame}
}
\description{
A S4 class that represents a DistributedDataFrame
}

