% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ddfKMeans}
\alias{ddfKMeans}
\title{ML model KMeans}
\usage{
ddfKMeans(x, ...)
}
\arguments{
\item{x}{a Distributed Data Frame.}

\item{centers}{number of cluster to be clustered.}

\item{runs}{number of runs.}

\item{maxIters}{max times of iterations.}
}
\value{
a IModel of distributedDataFrame for KMeans clustering
}
\description{
Return Returns IModel of distributedDataFrame
}

