% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DDFManager-class}
\alias{DDFManager}
\alias{DDFManager-class}
\title{An S4 class that represents a DistributedDataFrame Manager}
\usage{
DDFManager(engine = "spark")
}
\arguments{
\item{engine}{the underlying engine, for example: spark or local}
}
\description{
An S4 class that represents a DistributedDataFrame Manager
}

