% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R, R/spark-ddf.R
\docType{methods}
\name{ml.linear.regression}
\alias{ml.linear.regression}
\alias{ml.linear.regression,SparkDDF-method}
\title{ML Linear Regression}
\usage{
ml.linear.regression(x, ...)

\S4method{ml.linear.regression}{SparkDDF}(x, numIterations = 10L,
  stepSize = 1, miniBatchFraction = 1)
}
\arguments{
\item{x}{a Distributed Data Frame.}

\item{centers}{number of cluster to be clustered.}

\item{runs}{number of runs.}

\item{maxIters}{max times of iterations.}
}
\value{
a KMeanModel object
}
\description{
Train a Linear Regression algorithm

Linear Regression using Spark mllib's LinearRegressionSGD
}

