% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict}
\alias{predict}
\title{Predict the result of a sample using this ML model}
\usage{
predict(m, ...)
}
\arguments{
\item{m}{a MLModel}

\item{data}{the candidate sample data to be predicted,vector is expected}
}
\value{
predict result, class tag for classification, e.g.
}
\description{
Predict the result of a sample using this ML model
}

