% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sql2ddf}
\alias{sql2ddf}
\title{Execute a HiveQL and return a DistributedDataFrame}
\usage{
sql2ddf(x, sql, ...)
}
\arguments{
\item{x}{a DDFManager object}

\item{sql}{the query, only support SELECT ones}
}
\value{
a DistributedDataFrame
}
\description{
Execute a HiveQL and return a DistributedDataFrame
}

